job_matcher_crewAi/pyproject.toml
[tool.poetry]
name = "resume-job-matcher"
version = "0.2.0"
description = "AI system to match resumes to job descriptions using LLM and MongoDB"
authors = ["Wasseem Emad"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "0.1.24"
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.0"
openai = "^1.30.1"
pymongo = "^4.7.0"
spacy = "^3.7.4"
PyMuPDF = "^1.23.22"
requests = "^2.31.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
streamlit = "^1.46.0"


[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"